feat: task 207, 211: scene transition connected, stage ending condition connected
