+++ SAVE_THE_HAMSTER/Assets/Prefabs/Lava_Flowing_Shader/Materials/mtl_lava_simple.mat
@@ -27,7 +27,7 @@ Material:
     - _LavaTex:
         m_Texture: {fileID: 2800000, guid: 93a726948508cc14eab67fb1abe2daa8, type: 3}
         m_Scale: {x: 3, y: 1}
-        m_Offset: {x: 4.368914, y: 0}
+        m_Offset: {x: 2.3287435, y: 0}
     - _MainTex:
         m_Texture: {fileID: 2800000, guid: 348fa4d278c53ad4dacd07df1e63a248, type: 3}
         m_Scale: {x: 3, y: 1}
