+++ SAVE_THE_HAMSTER/Assets/Scripts/Collision.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a3657630aac9e1345b6c27dfd86d60c0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
