+++ SAVE_THE_HAMSTER/Assets/Scripts/Collision/AlmondCollision.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2db0ce0c9f946694dbea6513a85db9cc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
