+++ SAVE_THE_HAMSTER/Assets/Scripts/Collision/HamsterCollision.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2cb73c85223dc456c80471d9218009db
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
