+++ SAVE_THE_HAMSTER/Assets/Scripts/Collision/LavaCollision.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8a684c929de2ff7469b7462c1b2cd566
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
