--- SAVE_THE_HAMSTER/Assets/Scripts/CollisionDetection.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a68887af4a7bb4ab2bf9c39d4d52384a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
