--- SAVE_THE_HAMSTER/Assets/Scripts/Stage1Manager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f0ecd6e9176a7214082975c6b1ada46f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
