--- SAVE_THE_HAMSTER/Assets/Scripts/Stage3Manager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: df45589fba1d3445a9bfd7be11745460
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
