+++ SAVE_THE_HAMSTER/Assets/Scripts/StageManager.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0238c7a1a30953c47a0f66feffc9c3ab
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
