+++ SAVE_THE_HAMSTER/Assets/Scripts/StageManager/Stage2Manager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4c101d2921f2d495793cb943a7731dfa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
