+++ SAVE_THE_HAMSTER/Assets/Scripts/StageManager/StageManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 318ac1eccc240f84a8d879152d1d97a7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
