--- SAVE_THE_HAMSTER/Assets/Scripts/StickyBallCollision.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 29ae7b1f0a1c4ba4f91c54bd81438b5e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
