+++ SAVE_THE_HAMSTER/ProjectSettings/EditorBuildSettings.asset
@@ -17,10 +17,10 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/LoginScene.unity
     guid: 2ae7ad6d0cf969747a67496e68e4ebd0
-  - enabled: 1
+  - enabled: 0
     path: Assets/Scenes/PostStageScene.unity
     guid: 033722d8bd96faa44b16a7c8a16015b5
-  - enabled: 1
+  - enabled: 0
     path: Assets/Scenes/Stage1Scene_forBackUp.unity
     guid: 0f8b39ebec34d0d43a51db2439edee75
   - enabled: 1
@@ -31,5 +31,8 @@ EditorBuildSettings:
     guid: 45588f1aac8ef144bb7889618a7d39b4
   - enabled: 1
     path: Assets/Scenes/Stage1Scene.unity
-    guid: 7d6823e12555cac4e9f5fc7ef0c66322
+    guid: 2c98358d75c47e8408380bf89564173b
+  - enabled: 1
+    path: Assets/Scenes/Stage2Scene.unity
+    guid: 11b6ddf0e0fc9f642a045d4fdddd8c46
   m_configObjects: {}
